+++ inst/doc/cNORM-Demo.R
@@ -7,56 +7,45 @@ library(cNORM)
 library(ggplot2)
 
 
-## -----------------------------------------------------------------------------
+## ----fig0, fig.height = 4, fig.width = 7--------------------------------------
 library(cNORM)
-# Displays the first lines of the data of the example dataset 'elfe'
-head(elfe)
+# Models the 'raw' variable as a function of the discrete 'group' variable
+model <- cnorm(raw = elfe$raw, group = elfe$group)
 
 ## -----------------------------------------------------------------------------
-# Display some descriptive results by group
-by(elfe$raw, elfe$group, summary)
+print(model)
 
-## ----fig0, fig.height = 6, fig.width = 7--------------------------------------
-# Convenience method that does everything at once
-model <- cnorm(raw=elfe$raw, group=elfe$group)
-
-## ----fig1, fig.height = 6, fig.width = 7--------------------------------------
-# plot the information function per number of predictor
-plot(model, "subset", type = 0) # R2
-
-## ----fig2, fig.height = 6, fig.width = 7--------------------------------------
-plot(model, "subset", type = 3) # RMSE
+## ----fig1, fig.height = 4, fig.width = 7--------------------------------------
+    plot(model, "subset", type = 0)
 
 ## -----------------------------------------------------------------------------
-model <- cnorm(elfe$raw, elfe$group, terms = 5, plot = FALSE)
-
-# Search for intersecting percentiles
-checkConsistency(model, minNorm = 25, maxNorm = 75)
-
-## -----------------------------------------------------------------------------
-# Search for intersecting percentiles
-checkConsistency(model, minNorm = 25, maxNorm = 75)
-
-## ----fig4, fig.height = 6, fig.width = 7--------------------------------------
-plot(model, "norm")
-
-## ----fig5, fig.height = 6, fig.width = 7--------------------------------------
-# The plot can be split by group as well:
-plot(model, "norm", group = TRUE)    # specifies the grouping variable
-
-## ----fig6, fig.height = 6, fig.width = 7--------------------------------------
-plot(model, "derivative", minAge=1, maxAge=6, minNorm=20, maxNorm=80)
-# if parameters on age an norm are not specified, cnorm plots within
-# the ranges of the current dataset
-
-## -----------------------------------------------------------------------------
-# Predict norm score for raw score 15 and age 4.7
 predictNorm(15, 4.7, model, minNorm = 25, maxNorm = 75)
 
-# Predict raw score for normal score 55 and age 4.5
-predictRaw(55, 4.5, model)
 
-# ... or vectors, if you like ...
-predictRaw(c(40, 45, 50, 55, 60), c(2.5, 3, 3.5, 4, 4.5), model)
+## -----------------------------------------------------------------------------
+predictRaw(55, 4.5, model, minRaw = 0, maxRaw = 28)
 
+# ... or for several norm scores and age levels ...
+predictRaw(c(45, 50, 55), c(2.5, 3, 3.5), model) 
+
+## -----------------------------------------------------------------------------
+normTable(3, model, minRaw = 0, maxRaw = 28, minNorm=30.5, maxNorm=69.5, step = 1)
+
+## -----------------------------------------------------------------------------
+rawTable(3.5, model, minRaw = 0, maxRaw = 28, minNorm = 25, maxNorm = 75, step = 1, CI = .95, reliability = .89)
+
+# generate several raw tables
+table <- rawTable(c(2.5, 3.5, 4.5), model, minRaw = 0, maxRaw = 28)
+
+## ----fig2, fig.height = 7, fig.width = 7--------------------------------------
+plot(model, "raw", group = TRUE)
+
+## ----fig3, fig.height = 7, fig.width = 7--------------------------------------
+plot(model, "norm", group = TRUE, minNorm = 25, maxNorm = 75)
+
+## ----fig4, fig.height = 4, fig.width = 7--------------------------------------
+plot(model, "density", group = c (2, 3, 4))
+
+## ----fig5, fig.height = 4, fig.width = 7--------------------------------------
+plot(model, "derivative")